api/requirements.txt
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.11.0
Flask==2.0.3
Flask-Cors==5.0.0
fonttools==4.58.4
itsdangerous==2.2.0
Jinja2==3.1.1
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==1.26.4
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.4.0
scipy==1.16.0
six==1.17.0
threadpoolctl==3.6.0
Werkzeug==2.0.3